Load data from P78 service (#1191)
* Reverted removed loadDataFromP78() code

* Use SCENARIO from testDataForEachMethod MAP

+++ eceapp/src/main/java/com/autodesk/eceapp/testbase/ece/QuoteOrderTestBase.java
@@ -1,9 +1,12 @@
 package com.autodesk.eceapp.testbase.ece;
 
-import static com.autodesk.testinghub.core.common.EISTestBase.driver;
 import com.autodesk.eceapp.constants.BICECEConstants;
+import com.autodesk.eceapp.dto.QuoteDetails;
 import com.autodesk.eceapp.testbase.EceBICTestBase;
+import com.autodesk.eceapp.testbase.EceBICTestBase.Names;
 import com.autodesk.eceapp.testbase.EcePortalTestBase;
+import com.autodesk.eceapp.testbase.ece.DatastoreClient.OrderData;
+import com.autodesk.eceapp.testbase.ece.DatastoreClient.OrderFilters;
 import com.autodesk.eceapp.utilities.Address;
 import com.autodesk.testinghub.core.exception.MetadataException;
 import com.autodesk.testinghub.core.utils.ProtectedConfigFile;
@@ -12,8 +15,11 @@ import com.autodesk.testinghub.eseapp.constants.BICConstants;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.function.Consumer;
+import org.apache.commons.lang3.StringUtils;
 import org.openqa.selenium.JavascriptExecutor;
+import org.testng.util.Strings;
 
 public class QuoteOrderTestBase {
 
@@ -44,36 +50,55 @@ public class QuoteOrderTestBase {
     HashMap<String, String> results = new HashMap<String, String>();
     Address address = getBillingAddress(testDataForEachMethod);
 
-    Boolean isMultiLineItem =
-        !Objects.isNull(System.getProperty(BICECEConstants.IS_MULTILINE)) ? Boolean.valueOf(
-            System.getProperty(BICECEConstants.IS_MULTILINE)) : false;
+    Boolean isMultiLineItem = !Objects.isNull(System.getProperty(BICECEConstants.IS_MULTILINE)) ? Boolean.valueOf(
+        System.getProperty(BICECEConstants.IS_MULTILINE)) : false;
 
-    if (Objects.equals(System.getProperty(BICECEConstants.CREATE_PAYER), BICECEConstants.TRUE)) {
-      testResults = bicTestBase.createPayerAccount(testDataForEachMethod);
+    Boolean shouldPullFromDataStore =
+        !Objects.isNull(System.getProperty(BICECEConstants.PROJECT78_PULL_FLAG)) ? Boolean.valueOf(
+            System.getProperty(BICECEConstants.PROJECT78_PULL_FLAG)) : false;
+
+    String scenario = "";
+    if ((StringUtils.trimToNull(System.getProperty(BICECEConstants.SCENARIO)) != null)) {
+      scenario = System.getProperty(BICECEConstants.SCENARIO).trim();
+    } else if (StringUtils.trimToNull(testDataForEachMethod.get(BICECEConstants.SCENARIO)) != null) {
+      scenario = testDataForEachMethod.get(BICECEConstants.SCENARIO).trim();
     }
 
-    bicTestBase.goToDotcomSignin(testDataForEachMethod);
-    bicTestBase.createBICAccount(
-        new EceBICTestBase.Names(testDataForEachMethod.get(BICECEConstants.FIRSTNAME),
-            testDataForEachMethod.get(BICECEConstants.LASTNAME)),
-        testDataForEachMethod.get(BICECEConstants.emailid), PASSWORD, true);
+    if (shouldPullFromDataStore
+        && loadDataFromP78(BICECEConstants.QUOTE_TEST_NAME, scenario, testDataForEachMethod)) {
+      if (Strings.isNotNullAndNotEmpty(testDataForEachMethod.get(BICECEConstants.ADDRESS))) {
+        address = new Address(testDataForEachMethod.get(BICECEConstants.ADDRESS));
+      } else if (Strings.isNotNullAndNotEmpty(System.getProperty(BICECEConstants.ADDRESS))) {
+        address = new Address(System.getProperty(BICECEConstants.ADDRESS));
+      }
+      address.company = testDataForEachMethod.get("company");
+    } else {
+      if (Objects.equals(System.getProperty(BICECEConstants.CREATE_PAYER), BICECEConstants.TRUE)) {
+        testResults = bicTestBase.createPayerAccount(testDataForEachMethod);
+      }
+      bicTestBase.createBICAccount(
+          new Names(testDataForEachMethod.get(BICECEConstants.FIRSTNAME),
+              testDataForEachMethod.get(BICECEConstants.LASTNAME)),
+          testDataForEachMethod.get(BICECEConstants.emailid), PASSWORD, true);
 
-    String quoteId = pwsTestBase.createAndFinalizeQuote(address, testDataForEachMethod.get("quoteAgentCsnAccount"),
-        testDataForEachMethod.get("agentContactEmail"), testDataForEachMethod, isMultiLineItem);
+      String quoteId = pwsTestBase.createAndFinalizeQuote(address,
+          testDataForEachMethod.get("quoteAgentCsnAccount"),
+          testDataForEachMethod.get("agentContactEmail"), testDataForEachMethod, isMultiLineItem);
 
-    // Wait for Quote to sync from CPQ/SFDC to S4.
-    Util.printInfo("Keep calm, sleeping for 5min for Quote to sync to S4");
-    Util.sleep(300000);
+      // Wait for Quote to sync from CPQ/SFDC to S4.
+      Util.printInfo("Keep calm, sleeping for 5min for Quote to sync to S4");
+      Util.sleep(300000);
 
-    testDataForEachMethod.put(BICECEConstants.QUOTE_ID, quoteId);
-    testResults.put(BICECEConstants.QUOTE_ID, quoteId);
+      testDataForEachMethod.put(BICECEConstants.QUOTE_ID, quoteId);
+      testResults.put(BICECEConstants.QUOTE_ID, quoteId);
 
-    // Clear Storage data
-    JavascriptExecutor js = (JavascriptExecutor) driver;
-    js.executeScript("localStorage.clear();sessionStorage.clear();");
+      // Clear Storage data
+      JavascriptExecutor js = (JavascriptExecutor) bicTestBase.driver;
+      js.executeScript("localStorage.clear();sessionStorage.clear();");
 
-    // Signing out after quote creation
-    bicTestBase.getUrl(testDataForEachMethod.get("oxygenLogOut"));
+      // Signing out after quote creation
+      bicTestBase.getUrl(testDataForEachMethod.get("oxygenLogOut"));
+    }
 
     testDataForEachMethod.put("quote2OrderCartURL", bicTestBase.getQuote2OrderCartURL(
         testDataForEachMethod));
@@ -152,5 +177,42 @@ public class QuoteOrderTestBase {
   private Address getBillingAddress(LinkedHashMap<String, String> testData) {
     String billingAddress = getSerializedBillingAddress(testData);
     return new Address(billingAddress);
+  }
+
+  private boolean loadDataFromP78(String testName, String scenario,
+      LinkedHashMap<String, String> testDataForEachMethod) {
+    DatastoreClient dsClient = new DatastoreClient();
+    OrderFilters.OrderFiltersBuilder builder = OrderFilters.builder();
+
+    final String locale = Optional.ofNullable(testDataForEachMethod.get(BICECEConstants.LOCALE))
+        .orElseGet(() -> System.getProperty(BICECEConstants.LOCALE));
+
+    builder
+        .name(testName)
+        .locale(locale)
+        .scenario(scenario);
+
+    OrderData order = dsClient.grabOrder(builder.build());
+    try {
+      if (order == null || order.id == null) {
+        return false;
+      }
+      testDataForEachMethod.put(BICConstants.emailid, order.getEmailId());
+      testDataForEachMethod.put(BICECEConstants.QUOTE_ID, order.getQuoteId());
+      testDataForEachMethod.put("DS_ORDER_ID", order.getId().toString());
+
+      QuoteDetails quoteDetails = pwsTestBase.getQuoteDetails(
+          testDataForEachMethod.get("quoteAgentCsnAccount"), order.getQuoteId());
+
+      testDataForEachMethod.put("firstname", quoteDetails.getPurchaserFirstName());
+      testDataForEachMethod.put("lastname", quoteDetails.getPurchaserLastName());
+      testDataForEachMethod.put("company", quoteDetails.getEndCustomerName());
+      testDataForEachMethod.put("address", order.getAddress());
+
+    } catch (Exception e) {
+      Util.printInfo("Failed to fetch data from P78, for Quote Orders. Creating via PWS");
+      return false;
+    }
+    return true;
   }
 }
